+++ iotdb-core/node-commons/src/main/java/org/apache/iotdb/commons/pipe/connector/limiter/PipeEndPointRateLimiter.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.commons.pipe.connector.rateLimiter;
+package org.apache.iotdb.commons.pipe.connector.limiter;
 
 import com.google.common.util.concurrent.RateLimiter;
